packages/python-backend/pyproject.toml
[project]
name = "magicsync-backend"
version = "0.2.0"
description = "MagicSync Python FastAPI backend for RAG and AI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "langsmith>=0.1.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "markdown>=3.6",
    "duckduckgo-search>=7.0.0",
    "litellm>=1.55.0",
    "pyjwt>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
pdf-enhanced = [
    "pdfplumber>=0.11.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
target-version = "py311"
